requirements.txt
asn1crypto==0.22.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
cryptography==1.9
enum34==1.1.6
Flask==0.12.2
idna==2.6
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.10
keyring==10.4.0
keyrings.alt==2.2
MarkupSafe==1.0
mysql-connector-python==2.1.6
mysql-utilities==1.6.4
mysqlclient==1.3.10
paramiko==2.0.0
pexpect==4.2.1
pyasn1==0.1.9
pycrypto==2.6.1
pygobject==3.24.1
pyodbc==3.0.10
pysqlite==2.7.0
python-dateutil==2.6.1
pytz==2017.3
pyxdg==0.25
requests==2.18.4
requests-file==1.4.2
SecretStorage==2.3.1
six==1.11.0
tldextract==2.2.0
urllib3==1.22
Werkzeug==0.12.2